+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- DF-1104 Fixed migrations and other issues with using pdo_dblib driver and SQL Server as system database
 
 ## [0.10.0] - 2017-06-05
 ### Changed